auth/requirements.txt
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==43.0.0
defusedxml==0.7.1
Django==5.1
django-allauth==0.62.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-redis==5.4.0
django-rest-framework==0.1.0
django-rest-knox==5.0.1
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dynaconf==3.2.6
Faker==27.4.0
flower==1.2.0
frozenlist==1.4.1
humanize==4.10.0
hvac==2.3.0
idna==3.7
importlib==1.0.4
importlib_metadata==8.2.0
knox==0.1.14
kombu==5.4.0
loguru==0.7.2
multidict==6.0.5
oauthlib==3.2.2
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pycparser==2.22
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
redis==4.4.0
redis-cli==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
setuptools==72.2.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.0
tornado==6.4.1
twilio==9.2.3
tzdata==2024.1
urllib3==2.2.1
validators==0.33.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
zipp==3.20.0
